+++ src/tokenizer.h
@@ -1,4 +1,4 @@
 #include <cstring>
 #include <vector>
 
-void tokenizer(std::vector<std::string>& argList, std::string& input, int& argc);
+void tokenizer(std::vector<std::string>& argList, std::string& input, size_t& argc);
